// Manjeet_Pandey/frontend/src/api/models/modelTask.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * Task Manager API
 * A simple task management API built with Go and Gin.
 * OpenAPI spec version: 1.0
 */
import type { ModelTaskStatus } from "./modelTaskStatus";

export interface ModelTask {
  created_at?: string;
  description?: string;
  id?: string;
  name?: string;
  status?: ModelTaskStatus;
  updated_at?: string;
}